requirements.txt
torch>=2.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
openpyxl>=3.1
fastapi>=0.100
uvicorn>=0.23
psycopg2-binary>=2.9
httpx>=0.24
pydantic>=2.0
